+++ src/main/java/org/example/fibonacci/FibonacciStrategy.java
@@ -1,7 +1,8 @@
 package org.example.fibonacci;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface FibonacciStrategy {
-    BigInteger solveFibonacci(int n);
+    BigDecimal solveFibonacci(int n);
 }
